python/square_service/requirements.txt
fastapi
uvicorn
squareup
python-dotenv
requests